Fix FPM request_terminate_timeout not working
With HEROKU_PHP_GRACEFUL_SIGTERM=1, FPM children ignore SIGTERM (as sent by e.g. Heroku's process supervisor to all processes) so that the boot scripts can perform a clean shutdown by issuing SIGQUIT.

However, FPM internally depends on SIGTERM for terminating requests that exceed the timeout, so that's now being ignored.

We can, however, use SIGINT in this case, as it serves the same purpose.

Tests adapted to cover this case (the termination did previously work in some cases, e.g. when the script simply does a `sleep()`, as that is still interrupted by an incoming signal, even if the signal is being ignored).

GUS-W-14850175